Machine Learning Model: Regression Analysis
This assessment focuses on implementing a regression analysis ML model using the CRISP-DM method. The model seeks to determine which features are the best quality red wine indicators and generate insights into each of these factors to the model’s red wine quality. Dataset is based on red wine quality.